[REFACTOR] 도메인 전반에 걸친 Clock 주입 잔여분 반영 (Phase 1)
- MatchAttendanceService, OcrService, TeamRankingService 등 잔여 도메인 Clock 주입 완료
- API 응답 DTO 및 엔티티 내 시간 생성 로직을 결정론적 Clock 기반으로 통일
- 시스템 전역의 LocalDateTime.now() 암시적 호출 완전 제거
