+++ src/main/java/com/sparta/spartatigers/domain/liveboard/WeatherQueryController.java
@@ -6,7 +6,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.sparta.spartatigers.domain.auth.model.TokenClaim;
 import com.sparta.spartatigers.domain.liveboard.dto.MatchWeatherResponse;
+import com.sparta.spartatigers.global.aop.Auth;
 
 import lombok.RequiredArgsConstructor;
 
@@ -30,7 +32,7 @@ public class WeatherQueryController {
      * matchId로 경기 조회 → 연결된 구장의 현재 날씨와 시간대별 예보를 반환.
      */
     @GetMapping("/{matchId}/weather")
-    public ResponseEntity<MatchWeatherResponse> getMatchWeather(@PathVariable Long matchId) {
+    public ResponseEntity<MatchWeatherResponse> getMatchWeather(@PathVariable Long matchId, @Auth TokenClaim token) {
         return ResponseEntity.ok(weatherQueryService.getMatchWeather(matchId));
     }
 }
