--- src/main/java/com/sparta/spartatigers/domain/liveboardroom/controller/LiveBoardRoomController.java
@@ -1,56 +1,0 @@
-package com.sparta.spartatigers.domain.liveboardroom.controller;
-
-import java.time.LocalDate;
-import java.util.List;
-
-import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.sparta.spartatigers.domain.liveboardroom.dto.LiveBoardRoomResponseDto;
-import com.sparta.spartatigers.domain.liveboardroom.service.LiveboardRoomService;
-import com.sparta.spartatigers.global.response.ApiResponse;
-
-import lombok.RequiredArgsConstructor;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/liveboard/room")
-public class LiveBoardRoomController {
-
-	private final LiveboardRoomService liveboardRoomService;
-
-	@PostMapping
-	public ApiResponse<String> createRoomsForDay(
-		@RequestParam (required = false)
-		@DateTimeFormat(pattern = "yyyyMMdd") LocalDate anyday) {
-		if (anyday == null) {
-			anyday = LocalDate.now();
-		}
-		return ApiResponse.created(liveboardRoomService.createRoomsForDay(anyday));
-	}
-
-	@GetMapping
-	public List<LiveBoardRoomResponseDto> getRoomsForDay(
-		@RequestParam (required = false)
-		@DateTimeFormat(pattern = "yyyyMMdd") LocalDate anyday) {
-		if (anyday == null) {
-			anyday = LocalDate.now();
-		}
-		return liveboardRoomService.getRoomsForDay(anyday);
-	}
-
-	@DeleteMapping
-	public ApiResponse<String> deleteRoomsForDay(
-		@RequestParam (required = false)
-		@DateTimeFormat(pattern = "yyyyMMdd") LocalDate anyday) {
-		if (anyday == null) {
-			anyday = LocalDate.now();
-		}
-		return ApiResponse.success(liveboardRoomService.deleteRoomsForDay(anyday));
-	}
-}
