--- src/main/java/com/sparta/spartatigers/domain/liveboardroom/model/LiveBoardStatus.java
@@ -1,7 +1,0 @@
-package com.sparta.spartatigers.domain.liveboardroom.model;
-
-public enum LiveBoardStatus {
-	TODAY,
-	PAST,
-	UPCOMING
-}
