--- src/main/java/com/sparta/spartatigers/domain/match/model/LeagueType.java
@@ -1,7 +1,0 @@
-package com.sparta.spartatigers.domain.match.model;
-
-public enum LeagueType {
-	PRESEASON,
-	REGULAR,
-	POST_SEASON
-}
