--- src/main/java/com/sparta/spartatigers/domain/match/model/Match.java
@@ -1,72 +1,0 @@
-package com.sparta.spartatigers.domain.match.model;
-
-import java.time.LocalDateTime;
-
-import com.sparta.spartatigers.domain.common.entity.BaseEntity;
-import com.sparta.spartatigers.domain.team.model.Stadium;
-import com.sparta.spartatigers.domain.team.model.Team;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Table(name = "matches")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@SuperBuilder
-public class Match extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "match_id")
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private LeagueType leagueType;
-
-    @Column(nullable = false)
-    private int seasonYear;
-
-    @Column private LocalDateTime matchTime;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "home_team_id", nullable = false)
-    private Team homeTeam;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "away_team_id", nullable = false)
-    private Team awayTeam;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "stadium_id")
-    private Stadium stadium;
-
-    @Column
-    @Enumerated(value = EnumType.STRING)
-    private MatchResult matchResult;
-
-    @Column private Integer homeScore;
-
-    @Column
-    private Integer awayScore;
-
-    @Column private String remark; // 비고
-
-
-
-}
