--- src/main/java/com/sparta/spartatigers/domain/match/model/MatchResult.java
@@ -1,9 +1,0 @@
-package com.sparta.spartatigers.domain.match.model;
-
-public enum MatchResult {
-    HOME_WIN,
-    AWAY_WIN,
-    DRAW,
-    CANCEL,
-    NOT_PLAYED
-}
