+++ src/main/java/com/sparta/spartatigers/domain/matchAttendance/controller/OcrTestController.java
@@ -1,4 +1,4 @@
-package com.sparta.spartatigers.domain.matchAttendance.controller;
+package com.sparta.spartatigers.domain.liveboard.matchAttendance.controller;
 
 import java.io.IOException;
 
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.sparta.spartatigers.domain.matchAttendance.service.OcrService;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.service.OcrService;
 
 import lombok.RequiredArgsConstructor;
