+++ src/main/java/com/sparta/spartatigers/domain/matchAttendance/dto/MatchAttendanceUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.sparta.spartatigers.domain.matchAttendance.dto;
+package com.sparta.spartatigers.domain.liveboard.matchAttendance.dto;
 
 import java.util.List;
