+++ src/main/java/com/sparta/spartatigers/domain/matchAttendance/model/MatchAttendance.java
@@ -1,10 +1,10 @@
-package com.sparta.spartatigers.domain.matchAttendance.model;
+package com.sparta.spartatigers.domain.liveboard.matchAttendance.model;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import com.sparta.spartatigers.domain.common.entity.BaseEntity;
-import com.sparta.spartatigers.domain.match.model.Match;
+import com.sparta.spartatigers.domain.liveboard.match.model.Match;
 import com.sparta.spartatigers.domain.user.model.User;
 
 import jakarta.persistence.CascadeType;
