+++ src/main/java/com/sparta/spartatigers/domain/matchAttendance/repository/MatchAttendanceRepository.java
@@ -1,4 +1,4 @@
-package com.sparta.spartatigers.domain.matchAttendance.repository;
+package com.sparta.spartatigers.domain.liveboard.matchAttendance.repository;
 
 import java.util.List;
 
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sparta.spartatigers.domain.matchAttendance.model.MatchAttendance;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.model.MatchAttendance;
 
 public interface MatchAttendanceRepository extends JpaRepository<MatchAttendance, Long> {
