+++ src/main/java/com/sparta/spartatigers/domain/matchAttendance/service/MatchAttendanceService.java
@@ -1,4 +1,4 @@
-package com.sparta.spartatigers.domain.matchAttendance.service;
+package com.sparta.spartatigers.domain.liveboard.matchAttendance.service;
 
 import java.util.List;
 
@@ -12,15 +12,15 @@ import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 import com.sparta.spartatigers.domain.image.service.ImageStorageService;
-import com.sparta.spartatigers.domain.match.model.Match;
-import com.sparta.spartatigers.domain.match.repository.MatchRepository;
-import com.sparta.spartatigers.domain.matchAttendance.dto.MatchAttendanceRequestDto;
-import com.sparta.spartatigers.domain.matchAttendance.dto.MatchAttendanceResponseDto;
-import com.sparta.spartatigers.domain.matchAttendance.dto.MatchAttendanceUpdateRequestDto;
-import com.sparta.spartatigers.domain.matchAttendance.model.AttendanceImage;
-import com.sparta.spartatigers.domain.matchAttendance.model.AttendanceImageType;
-import com.sparta.spartatigers.domain.matchAttendance.model.MatchAttendance;
-import com.sparta.spartatigers.domain.matchAttendance.repository.MatchAttendanceRepository;
+import com.sparta.spartatigers.domain.liveboard.match.model.Match;
+import com.sparta.spartatigers.domain.liveboard.match.repository.MatchRepository;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.dto.MatchAttendanceRequestDto;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.dto.MatchAttendanceResponseDto;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.dto.MatchAttendanceUpdateRequestDto;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.model.AttendanceImage;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.model.AttendanceImageType;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.model.MatchAttendance;
+import com.sparta.spartatigers.domain.liveboard.matchAttendance.repository.MatchAttendanceRepository;
 import com.sparta.spartatigers.domain.user.model.User;
 import com.sparta.spartatigers.domain.user.repository.UserRepository;
 import com.sparta.spartatigers.global.exception.enums.ExceptionCode;
