+++ src/main/java/com/sparta/spartatigers/domain/ranking/dto/TeamRankingResponseDto.java
@@ -1,6 +1,6 @@
 package com.sparta.spartatigers.domain.ranking.dto;
 
-import com.sparta.spartatigers.domain.match.model.LeagueType;
+import com.sparta.spartatigers.domain.liveboard.match.model.LeagueType;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
