+++ src/main/java/com/sparta/spartatigers/domain/ranking/repository/TeamRankingRepositoryCustom.java
@@ -3,8 +3,8 @@ package com.sparta.spartatigers.domain.ranking.repository;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.sparta.spartatigers.domain.match.model.Match;
-import com.sparta.spartatigers.domain.match.model.LeagueType;
+import com.sparta.spartatigers.domain.liveboard.match.model.Match;
+import com.sparta.spartatigers.domain.liveboard.match.model.LeagueType;
 import com.sparta.spartatigers.domain.ranking.dto.TeamRankingStat;
 
 public interface TeamRankingRepositoryCustom {
